pyproject.toml
[project]
name = "kiddo"
version = "0.1.0"
description = "Event processing and calendar integration system"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
]

gmail = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
]

llm = [
    "openai>=1.0.0",
    "anthropic>=0.3.0",
]

calendar = [
    "icalendar>=5.0.0",
]

notifications = [
    "twilio>=8.0.0",
    "sendgrid>=6.0.0",
]

queue = [
    "redis>=4.5.0",
    "pika>=1.3.0",
    "kafka-python>=2.0.0",
]

database = [
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
]

all = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "openai>=1.0.0",
    "anthropic>=0.3.0",
    "icalendar>=5.0.0",
    "twilio>=8.0.0",
    "sendgrid>=6.0.0",
    "redis>=4.5.0",
    "pika>=1.3.0",
    "kafka-python>=2.0.0",
    "sqlalchemy>=2.0.0",
    "asyncpg>=0.28.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
markers = [
    "integration: Integration tests that require external services",
    "unit: Unit tests",
]
filterwarnings = [
    "ignore::DeprecationWarning:httplib2.*",
]

